Finish conflict tracking table basic functions
Conflict tracking table (CTT) is responsible for storing
proactive and pending tasks with corresponding states.
One task's state is defined as the its current executing
transaction's number, which is assigned by pre-processor using
__COUNTER__ marco.

CTT is separated to each parts:
table -> entry -> node
A CTT node is the basic element in CTT. It contains one task's
information, including pointer to coroutine storage, pointer to
coroutine's transaction descriptor and the state of this task.

A CTT entry contains multiple CTT nodes with linked list structure.
There are constant CTT entries in CTT, which is the maximum state
number + 1, with the last CTT entry is for garbage collection.