Finish Contention Intensity(CI) implemenation
To dynamic detect the contention among execution,
we used a metric called contention intensity(CI) to evaluate it.

CI detector is implemented in each thread and updates everytime
a transaction end(commit or abort).
There are still some hyper parameters to be well-tuned.

* Hyper Parameters:
    * ci_alpha:
        control the variation speed of CI, we prefered larger
        value.
    * CI_THRESHOLD:
        a threshold to determine if it is contention based on CI

The next step is to observe the statistic data from STAMP and
pick up a proper values for the below parameters.
Besides, the current CTT is the meaning of Pending Task Table(PTT)
in our paper. It should be modified in some day.